refactor: removed scheduled tasks
