+++ src/internal/index.ts
@@ -10,5 +10,4 @@ export * from './mathInterfaces';
 export { default as merge } from './merge';
 export { default as OrderedSet } from './OrderedSet';
 export * from './utils';
-export * from './worker';
 export * from './accessor';
