+++ src/internal/math.ts
@@ -1,6 +1,5 @@
 import type { ICategory, IndicesArray, UIntTypedArray } from '../model';
 import { IForEachAble, ISequence, isIndicesAble } from './interable';
-import { createWorkerCodeBlob, IPoorManWorkerScope, toFunctionBody } from './worker';
 import type {
   IAdvancedBoxPlotData,
   IStatistics,
@@ -1046,223 +1045,4 @@ export function categoricalValueCacheBuilder(length: number, categories: { name:
  */
 export function categoricalValueCache2Value<T extends { name: string }>(v: number, categories: T[]) {
   return v === 0 ? null : categories[v - 1];
-}
-
-function sortWorkerMain() {
-  // eslint-disable-next-line no-restricted-globals
-  const workerSelf = self as any as IPoorManWorkerScope;
-
-  // stored refs to avoid duplicate copy
-  const refs = new Map<string, UIntTypedArray | Float32Array | Int32Array | Float64Array | readonly string[]>();
-
-  const sort = (r: ISortMessageRequest) => {
-    if (r.sortOrders) {
-      sortComplex(r.indices, r.sortOrders);
-    }
-    const order = r.indices;
-
-    workerSelf.postMessage(
-      {
-        type: r.type,
-        uid: r.uid,
-        ref: r.ref,
-        order,
-      } as ISortMessageResponse,
-      [r.indices.buffer]
-    );
-  };
-
-  const setRef = (r: ISetRefMessageRequest) => {
-    if (r.data) {
-      refs.set(r.ref, r.data);
-    } else {
-      refs.delete(r.ref);
-    }
-  };
-
-  const deleteRef = (r: IDeleteRefMessageRequest) => {
-    if (!r.startsWith) {
-      refs.delete(r.ref);
-      return;
-    }
-
-    for (const key of Array.from(refs.keys())) {
-      if (key.startsWith(r.ref)) {
-        refs.delete(key);
-      }
-    }
-  };
-
-  const resolveRefs = <T extends UIntTypedArray | Float32Array | Int32Array | readonly string[]>(
-    r: IStatsWorkerMessage
-  ) => {
-    // resolve refs or save the new data
-
-    const data: T = r.data ? (r.data as any as T) : (refs.get(r.refData)! as any as T);
-    const indices = r.indices ? r.indices : r.refIndices ? (refs.get(r.refIndices)! as UIntTypedArray) : undefined;
-    if (r.refData) {
-      refs.set(r.refData, data);
-    }
-    if (r.refIndices) {
-      refs.set(r.refIndices, indices!);
-    }
-    return { data, indices };
-  };
-
-  const dateStats = (r: IDateStatsMessageRequest) => {
-    const { data, indices } = resolveRefs<Int32Array>(r);
-
-    const b = dateStatsBuilder(r.template);
-    if (indices) {
-      for (let ii = 0; ii < indices.length; ++ii) {
-        const v = data[indices[ii]];
-        b.push(dateValueCache2Value(v));
-      }
-    } else {
-      for (let i = 0; i < data.length; ++i) {
-        b.push(dateValueCache2Value(data[i]));
-      }
-    }
-    workerSelf.postMessage({
-      type: r.type,
-      uid: r.uid,
-      stats: b.build(),
-    } as IDateStatsMessageResponse);
-  };
-
-  const categoricalStats = (r: ICategoricalStatsMessageRequest) => {
-    const { data, indices } = resolveRefs<UIntTypedArray>(r);
-
-    const cats = r.categories.map((name) => ({ name }));
-    const b = categoricalStatsBuilder(cats);
-    if (indices) {
-      for (let ii = 0; ii < indices.length; ++ii) {
-        b.push(categoricalValueCache2Value(data[indices[ii]], cats));
-      }
-    } else {
-      for (let i = 0; i < data.length; ++i) {
-        b.push(categoricalValueCache2Value(data[i], cats));
-      }
-    }
-
-    workerSelf.postMessage({
-      type: r.type,
-      uid: r.uid,
-      stats: b.build(),
-    } as ICategoricalStatsMessageResponse);
-  };
-
-  const stringStats = (r: IStringStatsMessageRequest) => {
-    const { data, indices } = resolveRefs<readonly string[]>(r);
-
-    const b = stringStatsBuilder(r.topN ?? 10);
-    if (indices) {
-      for (let ii = 0; ii < indices.length; ++ii) {
-        b.push(data[indices[ii]]);
-      }
-    } else {
-      for (let i = 0; i < data.length; ++i) {
-        b.push(data[i]);
-      }
-    }
-
-    workerSelf.postMessage({
-      type: r.type,
-      uid: r.uid,
-      stats: b.build(),
-    } as IStringStatsMessageResponse);
-  };
-
-  const numberStats = (r: INumberStatsMessageRequest) => {
-    const { data, indices } = resolveRefs<Float32Array>(r);
-
-    const b = numberStatsBuilder(r.domain ?? [0, 1], r.numberOfBins);
-
-    if (indices) {
-      for (let ii = 0; ii < indices.length; ++ii) {
-        b.push(data[indices[ii]]);
-      }
-    } else {
-      for (let i = 0; i < data.length; ++i) {
-        b.push(data[i]);
-      }
-    }
-
-    workerSelf.postMessage({
-      type: r.type,
-      uid: r.uid,
-      stats: b.build(),
-    } as INumberStatsMessageResponse);
-  };
-
-  const boxplotStats = (r: IBoxPlotStatsMessageRequest) => {
-    const { data, indices } = resolveRefs<Float32Array>(r);
-
-    const b = boxplotBuilder(indices ? indices.length : undefined);
-
-    let stats: IAdvancedBoxPlotData;
-    if (!indices) {
-      stats = b.buildArr(data);
-    } else {
-      for (let ii = 0; ii < indices.length; ++ii) {
-        b.push(data[indices[ii]]);
-      }
-      stats = b.build();
-    }
-
-    workerSelf.postMessage({
-      type: r.type,
-      uid: r.uid,
-      stats,
-    } as IBoxPlotStatsMessageResponse);
-  };
-
-  // message type to handler function
-  const msgHandlers: { [key: string]: (r: any) => void } = {
-    sort,
-    setRef,
-    deleteRef,
-    dateStats,
-    categoricalStats,
-    numberStats,
-    boxplotStats,
-    stringStats,
-  };
-
-  workerSelf.addEventListener('message', (evt) => {
-    const r = evt.data;
-    if (typeof r.uid !== 'number' || typeof r.type !== 'string') {
-      return;
-    }
-    const h = msgHandlers[r.type];
-    if (h) {
-      h(r);
-    }
-  });
-}
-
-/**
- * copy source code of a worker and create a blob out of it
- * to avoid webpack imports all the code functions need to be in this file
- * @internal
- */
-export function createWorkerBlob() {
-  return createWorkerCodeBlob([
-    pushAll.toString(),
-    quantile.toString(),
-    numberStatsBuilder.toString(),
-    boxplotBuilder.toString(),
-    computeGranularity.toString(),
-    pushDateHist.toString(),
-    dateStatsBuilder.toString(),
-    categoricalStatsBuilder.toString(),
-    stringStatsBuilder.toString(),
-    createIndexArray.toString(),
-    asc.toString(),
-    desc.toString(),
-    sortComplex.toString(),
-    dateValueCache2Value.toString(),
-    categoricalValueCache2Value.toString(),
-    toFunctionBody(sortWorkerMain),
-  ]);
 }
