+++ src/provider/index.ts
@@ -1,8 +1,6 @@
 export { default as DataProvider } from './ADataProvider';
 export { default as LocalDataProvider } from './LocalDataProvider';
-export { default as RemoteDataProvider } from './RemoteDataProvider';
 export * from './LocalDataProvider';
-export * from './RemoteDataProvider';
 export { deriveColors, deriveColumnDescriptions, exportRanking, exportTable } from './utils';
 export * from './interfaces';
 export { tasksAll } from './tasks';
